extract post-processing API to library, enable full pipeline in cortex_worker
Move PostOptions and run_post_processing from latexml_oxide binary into
the library (latexml_oxide/src/post.rs) as public API. This allows both
latexml_oxide and cortex_worker to share the same post-processing code.

cortex_worker now runs the full CorTeX conversion pipeline matching the
Perl TexToHtmlWorker settings:
- Presentation MathML + Content MathML
- XSLT with LaTeXML-html5.xsl
- nodefaultresources, noinvisibletimes, mathtex
- Scan → Bibliography → CrossRef → Graphics → MathML → XSLT → HTML5

The latexml_oxide binary delegates to the same library function,
removing ~200 lines of duplicated code.
