+++ latexml_oxide/bin/cortex_worker.rs
@@ -182,8 +182,27 @@ impl LatexmlWorker {
       .result
       .ok_or_else(|| format!("Conversion failed for {}", main_tex))?;
 
-    // 4. Get log and status
-    // TODO: add post-processing (MathML + XSLT) once latexml_post APIs are public
+    // 4. Post-process: MathML + XSLT (matching CorTeX tex_to_html settings)
+    let xml = latexml::post::run_post_processing(&xml, &latexml::post::PostOptions {
+      pmml: self.profile.pmml,
+      cmml: true, // CorTeX produces both pmml and cmml
+      keep_xmath: false,
+      stylesheet: Some("resources/XSLT/LaTeXML-html5.xsl"),
+      destination: None,
+      source_directory: Some(&source_dir),
+      nodefaultresources: self.profile.nodefaultresources,
+      css_files: &[],
+      js_files: &[],
+      noinvisibletimes: self.profile.noinvisibletimes,
+      mathtex: self.profile.mathtex,
+      navigationtoc: None,
+      split: false,
+      split_xpath: None,
+      split_naming: None,
+      xslt_parameters: &[],
+    });
+
+    // 5. Get log and status
     let log = response.log;
     let status_str = format!("Status:conversion:{}", response.status_code);
