+++ latexml_oxide/src/lib.rs
@@ -4,4 +4,5 @@ extern crate latexml_package;
 pub mod converter;
 pub mod core_interface;
 pub mod ini_tex;
+pub mod post;
 pub mod util;
